requirements.txt
grpcio
grpcio-tools
opencv-python
pytest
